+++ ModManifestBuilder/AddSMAPIDependencyReferencesTask.cs
@@ -1,16 +1,11 @@
-using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Net;
-using System.Text;
 
 using Microsoft.Build.Framework;
 using Microsoft.Build.Utilities;
 
 using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 
 namespace Leclair.Stardew.ModManifestBuilder;
 
@@ -30,6 +25,11 @@ public class AddSMAPIDependencyReferencesTask : Task {
 	[Required]
 	public string? ModsPath { get; set; }
 
+	/// <summary>
+	/// This is the path to the project file being built.
+	/// </summary>
+	public string? ProjectPath { get; set; }
+
 	#endregion
 
 	#region Outputs
@@ -47,7 +47,7 @@ public class AddSMAPIDependencyReferencesTask : Task {
 		Dictionary<string, ITaskItem> references = Dependencies.Where(x => {
 			if (x.GetMetadata("Reference").TryParseBoolean(out bool val) && val) {
 				if (x.GetMetadata("Required").TryParseBoolean(out bool v) && !v)
-					Log.LogAddRef(LogLevel.Error, $"Dependency '{x.ItemSpec}' wants a reference, but is set to not required.");
+					Log.LogAddRef(LogLevel.Error, $"Dependency '{x.ItemSpec}' wants a reference, but is set to not required.", ProjectPath);
 
 				return true;
 			}
@@ -56,113 +56,78 @@ public class AddSMAPIDependencyReferencesTask : Task {
 		}).ToDictionary(x => x.ItemSpec, x => x);
 
 		if (references.Count == 0) {
-			Log.LogAddRef(LogLevel.Debug, "No <SMAPIDependency /> entries want references.");
+			Log.LogAddRef(LogLevel.Debug, "No <SMAPIDependency /> entries want references.", ProjectPath);
 			return true;
 		}
 
 		List<ITaskItem> result = new List<ITaskItem>();
 
 		if (string.IsNullOrEmpty(ModsPath) || ! Directory.Exists(ModsPath)) {
-			Log.LogAddRef(LogLevel.Error, $"The provided mods directory is empty or missing.");
+			Log.LogAddRef(LogLevel.Error, $"The provided mods directory is empty or missing.", ProjectPath);
 			return false;
 		}
 
-		RecursivelyFindMods(ModsPath!, references, result, false);
+		bool OnFound(string path, ModManifest manifest, SemanticVersion version) {
+			if (references.TryGetValue(manifest.UniqueID!, out var wanted)) {
+				references.Remove(manifest.UniqueID!);
+				string rawVersion = wanted.GetMetadata("Version");
+				SemanticVersion? wantedVersion = null;
 
-		foreach (var entry in references)
-			Log.LogAddRef(LogLevel.Error, $"Could not find mod '{entry.Key}' to reference.");
+				string? from = wanted.GetMetadata("DefiningProjectFullPath");
+				if (string.IsNullOrWhiteSpace(from))
+					from = ProjectPath!;
+
+				string? dllPath = string.IsNullOrEmpty(manifest.EntryDll)
+					? null
+					: Path.Combine(path, manifest.EntryDll!);
+
+				if (dllPath is null) {
+					Log.LogAddRef(LogLevel.Error, $"The mod '{manifest.UniqueID}' does not have an EntryDll and therefore cannot be used as a reference.", from);
+
+				} else if (!File.Exists(dllPath)) {
+					Log.LogAddRef(LogLevel.Error, $"Cannot find EntryDll '{manifest.EntryDll}' for mod '{manifest.UniqueID}'.", from);
+
+				} else if (rawVersion is not null && ! SemanticVersion.TryParse(rawVersion, out wantedVersion)) {
+					Log.LogAddRef(LogLevel.Error, $"Could not parse required version '{rawVersion}' for mod '{manifest.UniqueID}'.", from);
+
+				} else if (wantedVersion is not null && wantedVersion.IsNewerThan(version)) {
+					Log.LogAddRef(LogLevel.Error, $"Mod '{manifest.UniqueID}' is version '{version}', which does not meet required version '{wantedVersion}'.", from);
+
+				} else {
+					Log.LogAddRef(LogLevel.Info, $"Found mod '{manifest.UniqueID}' with version '{version}' at '{dllPath}'.", from);
+
+					string assembly = wanted.GetMetadata("Assembly");
+					if (string.IsNullOrEmpty(assembly))
+						assembly = Path.GetFileNameWithoutExtension(manifest.EntryDll);
+
+					result.Add(new TaskItem(assembly, new Dictionary<string, string> {
+						{ "Private", "False" },
+						{ "HintPath", dllPath },
+						{ "SMAPIDependency_VersionBehavior", wanted.GetMetadata("VersionBehavior") },
+						{ "DefiningProjectFullPath", wanted.GetMetadata("DefiningProjectFullPath") },
+						{ "DefiningProjectDirectory", wanted.GetMetadata("DefiningProjectDirectory") },
+						{ "DefiningProjectName", wanted.GetMetadata("DefiningProjectName") },
+						{ "DefiningProjectExtension", wanted.GetMetadata("DefiningProjectExtension") }
+					}));
+				}
+			}
+
+			// Keep iterating as long as we have references to find.
+			return references.Count > 0;
+		}
+
+		Utilities.RecursivelyFindMods(ModsPath!, OnFound, checkManifest: false);
+
+		foreach (var entry in references) {
+			string? from = entry.Value.GetMetadata("DefiningProjectFullPath");
+			if (string.IsNullOrWhiteSpace(from))
+				from = ProjectPath!;
+
+			Log.LogAddRef(LogLevel.Error, $"Could not find mod '{entry.Key}' to reference.", from);
+		}
 
 		ResolvedReferences = result.ToArray();
 
 		return !Log.HasLoggedErrors;
 	}
-
-	private void RecursivelyFindMods(string path, Dictionary<string, ITaskItem> targets, List<ITaskItem> found, bool checkManifest = true) {
-
-		string manifestFile = Path.Combine(path, "manifest.json");
-
-		// If there's a manifest here, don't recurse into it.
-		if (File.Exists(manifestFile)) {
-			if (checkManifest)
-				TryConsumeMod(path, targets, found);
-
-			return;
-		}
-
-		if (targets.Count == 0)
-			return;
-
-		foreach (string dir in Directory.EnumerateDirectories(path)) {
-			string reldir = Utilities.GetRelativePath(path, dir);
-
-			// SMAPI skips folders starting with a dot.
-			if (reldir.StartsWith("."))
-				continue;
-
-			RecursivelyFindMods(dir, targets, found);
-			if (targets.Count == 0)
-				return;
-		}
-	}
-
-	private void TryConsumeMod(string path, Dictionary<string, ITaskItem> targets, List<ITaskItem> found) {
-		string manifestFile = Path.Combine(path, "manifest.json");
-
-		if (!File.Exists(manifestFile))
-			return;
-
-		ModManifest manifest;
-		try {
-			manifest = JsonConvert.DeserializeObject<ModManifest>(File.ReadAllText(manifestFile));
-		} catch {
-			return;
-		}
-
-		if (manifest?.UniqueID is null || !targets.TryGetValue(manifest.UniqueID, out ITaskItem wanted))
-			return;
-
-		targets.Remove(manifest.UniqueID);
-
-		string? dllPath = string.IsNullOrEmpty(manifest.EntryDll)
-						? null
-						: Path.Combine(path, manifest.EntryDll!);
-
-		if (dllPath is null) {
-			Log.LogAddRef(LogLevel.Error, $"Mod '{manifest.UniqueID}' does not have an EntryDll and therefore cannot be used as a reference.");
-			return;
-		}
-
-		if (!File.Exists(dllPath)) {
-			Log.LogAddRef(LogLevel.Error, $"Cannot find EntryDll '{manifest.EntryDll}' for mod '{manifest.UniqueID}'.");
-			return;
-		}
-
-		if (!SemanticVersion.TryParse(manifest.Version, out var modVer) || modVer is null) {
-			Log.LogAddRef(LogLevel.Error, $"Could not parse version '{manifest.Version}' of mod '{manifest.UniqueID}'.");
-			return;
-		}
-
-		string? rawWantVer = wanted.GetMetadata("Version");
-		if (rawWantVer is null || !SemanticVersion.TryParse(rawWantVer, out var wantVer) || wantVer is null) {
-			Log.LogAddRef(LogLevel.Error, $"Could not parse required version '{rawWantVer}' for mod '{manifest.UniqueID}'.");
-			return;
-		}
-
-		if (wantVer.IsNewerThan(modVer)) {
-			Log.LogAddRef(LogLevel.Error, $"Mod '{manifest.UniqueID}' is version '{modVer}', which does not meet required version '{wantVer}'.");
-			return;
-		}
-
-		Log.LogAddRef(LogLevel.Info, $"Found mod '{manifest.UniqueID}' with version '{modVer}' at '{dllPath}'.");
-
-		string assembly = wanted.GetMetadata("Assembly");
-		if (string.IsNullOrEmpty(assembly))
-			assembly = Path.GetFileNameWithoutExtension(manifest.EntryDll);
-
-		found.Add(new TaskItem(assembly, new Dictionary<string, string> {
-			{ "Private", "False" },
-			{ "HintPath", dllPath }
-		}));
-	}
-
 }