+++ TestModTwo/ModEntry.cs
@@ -1,5 +1,7 @@
 using System;
 
+using Newtonsoft.Json;
+
 using StardewModdingAPI;
 
 namespace TestModTwo;
@@ -8,6 +10,9 @@ public class ModEntry : Mod {
 
 	public override void Entry(IModHelper helper) {
 		Monitor.Log("Loaded.", LogLevel.Info);
+
+		Monitor.Log(JsonConvert.SerializeObject(null), LogLevel.Info);
+
 	}
 
 }